# Conflicts:
#	packages/server/tests/contracts/BaseHTS.json

+++ packages/server/tests/contracts/BaseHTS.sol
@@ -9,9 +9,9 @@ contract BaseHTS is FeeHelper {
     string name = "tokenName";
     string symbol = "tokenSymbol";
     string memo = "memo";
-    uint initialTotalSupply = 1000;
-    uint32 maxSupply = 1000;
-    uint decimals = 8;
+    uint64 initialTotalSupply = 1000;
+    int64 maxSupply = 1000;
+    uint32 decimals = 8;
     bool freezeDefaultStatus = false;
 
     event CreatedToken(address tokenAddress);
